+++ flutter_modular/example/lib/app/app_module.dart
@@ -9,7 +9,10 @@ import 'modules/tabs/tabs_module.dart';
 
 class AppModule extends MainModule {
   @override
-  List<Bind> get binds => [];
+  List<Bind> get binds => [
+        $HomeRealController,
+      ];
+
 
   @override
   List<Router> get routers => [