+++ flutter_modular/example/lib/app/gentest/test.controller.dart
@@ -1,7 +1,12 @@
 import 'package:flutter_modular/flutter_modular.dart';
 
+import '../app_bloc.dart';
+
+
 part 'test.controller.g.dart';
 
-class TestController {
-  TestController();
+@Injectable()
+class HomeRealController {
+  final AppBloc bloc;
+  HomeRealController(this.bloc);
 }