+++ flutter_modular/lib/flutter_modular.dart
@@ -1,5 +1,6 @@
 library flutter_modular;
 
+export 'flutter_modular_annotations.dart';
 export 'src/exceptions/modular_error.dart';
 export 'src/inject/bind.dart';
 export 'src/inject/inject.dart';