+++ modular_codegen/CHANGELOG.md
@@ -1,3 +1,7 @@
-## [1.0.0] - 03/07/2020.
+## [1.0.1] - 13/07/2020.
 
-* Welcome Codegen!!!
+* Fix error
+
+## [1.0.0] - 12/07/2020.
+
+* Welcome Modular Codegen!!!